Fixed tidy-html5 installation failure and put cmake options in SB to save having to search through CMakeLists.txt to find out what are available. Removed tidy-html5-master.zip from src.
Added langcodes [a modified tde-i18n README] and included it in BUILD-TDE.sh as a README option for tde-i18n.

Changed 1st screen intro re: mirror selection by geoIP.

The if loop for setting the COMPILER_CXX variable can be done in one line, and will then be in the same format as the other variable exports.