Added cpu optimization
Added a section to BUILD-TDE.sh for cpu optimization.

It only functions for x86 - the build for any other cpu type works as before.

I would think that most people building these packages would be doing so for their own use on the machine where the build takes place, so it makes sense to provide an option to optimize the build for that machine.

Added the cpu option to get-source.sh, tqt3.SlackBuild, and tqca-tls.SlackBuild. Moved setting ARCH from get-source.sh and koffice.SlackBuild into BUILD-TDE.sh where it's used for this new option.

Added a summary of the selected options to the final dialog screen.

Other changes to BUILD-TDE.sh are an internal clean-up, mainly making use of dialog options.